replace atomic_lock with pthread_spinlock_t
Quote from Section 3.4.1.2 Spin-Wait and Idle Loops in itel x86
Optmiization Reference Manual

'The Pentium 4 processor introduces a new PAUSE instruction; the
instruction is architecturally a NOP on Intel 64 and IA-32 processor
implementations.
To the Pentium 4 and later processors, this instruction acts as a hint
that the code sequence is a spin-wait loop. Without a PAUSE instruction
in such loops, the Pentium 4 processor may suffer a severe penalty when
exiting the loop because the processor may detect a possible memory
order violation. Inserting the PAUSE instruction significantly reduces
the likelihood of a memory order violation and as a result improves
performance.'

And in Linux, the 'cpu_relax' function used in spinlock calls rep;nop
which is the same as PAUSE and say "REP NOP (PAUSE) is a good thing to
insert into busy-wait loops."