refactor core.lua
In Lua, `thread` is a basic concept used to implement `coroutine`,
the first thread is main thread, and others thread must be `coroutine`,
here, `coroutine` is used as basic 'schedule unit'.
For clarity, a new concept should be redefined so that it can be
distinguished from normal coroutine,
so, the 'schedule unit' is called `task`.

Invariants:
1. all the task must be resumed in main thread
2. all the logic must be wrapped by task
3. when the coroutine in `copool`, there muse be no object refer to it

BTW, coroutine pool can't reuse coroutine which throw exception, because it
may break Invariants.3, consider the following code:
`
core.fork(function()
	local co = core.running()
	core.timeout(1000, function()core.wakeup(co)end)
	do_something()
	core.wait()
end)
`
when do_something throw exception, the coroutine is over, but the timer
task still refer the task which is forked.
If reuse this coroutine at this time, it will break Invariants.3