replace redis:connect with redis.new
Using redis.new defers the initial connection until the first operation,
This change improves the caching of the database object.
Because when redis:connect is invoked in the main module and
the module is subsequently required by others, a yield error occurs.
If we want check the redis connection when call redis.new, we can call
redis:ping() manually.